requirements.txt
boto3==1.35.22
PyYAML==6.0.2

